fix: 그룹 게시물 서비스 (조회와 나머지로 서비스 분리)
